Big refactor to separate Entity from Report, now have IEP and transferIEP objects
also,

changed all data references to be refIds to serve as the basis for a query in-zone
added referenceObjects to serve as the query target out of zone, ie, at transfer destination

added specific refId types for each of the kinds of data pointed to (and carried in the iepTransfer/referenceObjects)

NOTE:

These two files reference SIFNAxSRECommonTypes.xsd. However, when an include is added, the overall model (collection.xsd)
doesn't compile. Unfortunately, that means that the iep and transferIep do not compile on their own.

NADM/Entity/SIFNAindividualizedEducationPlan/plan/iepPlanPresentLevelPerformanceFunctionalType.xsd
NADM/Entity/SIFNAindividualizedEducationPlan/eligibility/iepEligibilityIdeaDisabilityType.xsd

It is set so the model compiles and then local includes are commented out. I do not know how to fix this
so that, as with all the other includes, the redundant invocation doesn't cause a problem.